fix(openai-agents): address PR review comments
- Refactor APM plugins: switch from ClientPlugin/CompositePlugin object
  spread to TracingPlugin with individual static prefix/spanName per class;
  export as arrays
- Remove all model/provider/host/usage tags from APM spans (LLMObs-only)
- Extract LLMObs helpers into utils.js (getModelProvider, extractAgentManifest,
  extractInputMessages, extractOutputMessages, etc.) for testability
- Fix getModelProvider to fall back to 'unknown' instead of empty string
- Fix TypeScript definition comment for openai-agents integration
- Restore accidentally-dropped supported-configurations.json entries
- Add DD_TRACE_OPENAI_AGENTS_ENABLED to supported-configurations.json
- Fix test-setup.js: use versioned absolute paths for @openai/agents-openai
  and openai resolution; fix module.Agent → clientModule.Agent references
- Fix LLMObs spec: use withVersions() wrapper, fix openai require path,
  add metadata: MOCK_NOT_NULLISH assertions for run() workflow spans
