+++ packages/dd-trace/src/config/supported-configurations.json
@@ -533,6 +533,13 @@
         "default": "false"
       }
     ],
+    "DD_ENABLE_LAGE_PACKAGE_NAME": [
+      {
+        "implementation": "A",
+        "type": "boolean",
+        "default": "false"
+      }
+    ],
     "DD_CIVISIBILITY_MANUAL_API_ENABLED": [
       {
         "implementation": "A",
@@ -769,7 +776,9 @@
         "implementation": "B",
         "type": "boolean",
         "default": "true",
-        "internalPropertyName": "propagateProcessTags.enabled"
+        "configurationNames": [
+          "propagateProcessTags.enabled"
+        ]
       }
     ],
     "DD_EXPERIMENTAL_TEST_OPT_SETTINGS_CACHE": [
